Create Clean
Batch script that you can use to clear the desktop cache on a Windows machine. This script will delete the contents of the temporary files directory, which is often used as a cache by various applications.  Save this script with a .bat extension, and you can run it to clear the cache. 
Please note that this script will delete all files in the temp directories, so make sure there are no files you wish to keep before running it.